Make MP button link to prefilled private message form
Updated all instances of the 'MP' (private message) button to link directly to the new private message thread form with the recipient prefilled via a URL parameter. Added JavaScript to prefill the recipient field in the form based on the 'user' URL parameter. Also updated the README to mark this task as complete and added a new cleanup task for phpBB links.