Add beta chatbox in the classic frontend
Introduces a beta chatbox to the classic theme by adding new static images, a JavaScript chat client, and updating the index.html template to include the chatbox UI and related scripts. The chatbox is conditionally rendered based on a debug query parameter and is still expecting xooit files for the actual messages. This will get replaced.